Moved most smoke particle logic to its own class
